pyproject.toml
[tool.poetry]
name = "countrymark-sdk"
version = "0.1.0"
description = ""
authors = ["Rajdeep"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
sqlalchemy = "^2.0.31"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
